feat: Network가 끊겼을 때 Room을 활용하여 데이터 보여주는 작업

+++ app/src/main/java/com/choi/part3_ch08/data/repository/ContentRepositoryImpl.kt
@@ -9,7 +9,6 @@ import com.choi.part3_ch08.domain.model.Content
 import com.choi.part3_ch08.domain.repository.ContentRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import java.lang.Exception
 import javax.inject.Inject
 
 class ContentRepositoryImpl @Inject constructor(
@@ -21,8 +20,13 @@ class ContentRepositoryImpl @Inject constructor(
     override suspend fun save(item: Content): Boolean {
         // 네트 워크가 없을 때 예외처리를 하기 위해
         return try {
-            contentService.saveItem(item.toRequest())
-            contentDao.insert(item.toEntity())
+            contentService.saveItem(item.toRequest()).also {
+                if (it.success) {
+                    it.data?.let { contentDto ->
+                        contentDao.insert(contentDto.toEntity())
+                    }
+                }
+            }
             true
         } catch (e: Exception) {
             false
@@ -31,8 +35,13 @@ class ContentRepositoryImpl @Inject constructor(
 
     override suspend fun update(item: Content): Boolean {
         return try {
-            contentService.updateItem(item.toRequest())
-            contentDao.insert(item.toEntity())
+            contentService.updateItem(item.toRequest()).also {
+                if (it.success) {
+                    it.data?.let { contentDto ->
+                        contentDao.insert(contentDto.toEntity())
+                    }
+                }
+            }
             true
         } catch (e: Exception) {
             false
@@ -41,27 +50,31 @@ class ContentRepositoryImpl @Inject constructor(
 
     override fun loadList(): Flow<List<Content>> {
         return flow {
-//            contentDao.selectAll().collect { list ->
-//                emit(list.map { it.toContent() })
-//            }
-            emit(
-                try {
-                    contentService.getList().data.map { it.toContent() }
-                } catch (e: Exception) {
-                    emptyList()
+            try {
+                contentService.getList().data.also {list ->
+                    contentDao.insertAll(list.map { it.toEntity() })
                 }
-            )
+            } finally {
+                // finally 를 써서 내부 DB 에 있는 정보 불러 오기
+                contentDao.selectAll().collect { list ->
+                    emit(list.map { it.toContent() })
+                }
+            }
+
         }
     }
 
     override suspend fun delete(item: Content): Boolean {
         return try {
-            item?.id?.let {id->
-                contentService.deleteItem(id)
+            item?.id?.let { id ->
+                contentService.deleteItem(id).also {
+                    if (it.success) {
+                        contentDao.delete(item.toEntity())
+                    }
+                }
             }
-            contentDao.delete(item.toEntity())
             true
-        } catch (e:Exception) {
+        } catch (e: Exception) {
             false
         }
     }
